requirements.txt
python-dotenv
requests
openai>=1.40.0
ipython
numpy